// frontend/src/app/model/persona-request.ts
export interface PersonaRequest {
  idPersona: number;
  apellidoPaterno: string;
  apellidoMaterno: string;
  nombres: string;
  idSexo: string;
  fechaNacimiento: Date;
  idTipoDocumento: number;
  numDocumento: string;
  direccion: string;
  telefono: string;
  idUbigeo: string;
  username?: string;
}